# TSServer/src/main/java/de/hanke/arnim/TSServer/api/TimeseriesDefinitionByParametersApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.14).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package de.hanke.arnim.TSServer.api;

import de.hanke.arnim.TSServer.model.AccessParamterHeadDto;
import de.hanke.arnim.TSServer.model.DeleteParameterHeadDto;
import de.hanke.arnim.TSServer.model.Error;
import java.util.List;
import de.hanke.arnim.TSServer.model.PersistenceParameterHeadDto;
import de.hanke.arnim.TSServer.model.TimeSeriesHead;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import javax.validation.Valid;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-06-01T11:50:12.361Z")

@Api(value = "timeseriesDefinitionByParameters", description = "the timeseriesDefinitionByParameters API")
@RequestMapping(value = "/TSTool-api")
public interface TimeseriesDefinitionByParametersApi {

    @ApiOperation(value = "Removes timeseries heads. The Timeseries must exists!", nickname = "deleteTimeseriesDefinitionByDeleteParameterHeadDtos", notes = "", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK"),
        @ApiResponse(code = 401, message = "unauthorized", response = Error.class),
        @ApiResponse(code = 404, message = "resource not found", response = Error.class),
        @ApiResponse(code = 200, message = "error", response = Error.class) })
    @RequestMapping(value = "/timeseriesDefinitionByParameters",
        produces = { "application/json" }, 
        method = RequestMethod.DELETE)
    ResponseEntity<Void> deleteTimeseriesDefinitionByDeleteParameterHeadDtos(@ApiParam(value = ""  )  @Valid @RequestBody List<DeleteParameterHeadDto> deleteParametersHeadDto);


    @ApiOperation(value = "Get timeseries-Definitions for requested AccessParamterHeadDtos", nickname = "getTimeseriesDefinitionByAccessParamterHeadDtos", notes = "", response = TimeSeriesHead.class, responseContainer = "List", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = TimeSeriesHead.class, responseContainer = "List"),
        @ApiResponse(code = 401, message = "unauthorized", response = Error.class),
        @ApiResponse(code = 404, message = "resource not found", response = Error.class),
        @ApiResponse(code = 200, message = "error", response = Error.class) })
    @RequestMapping(value = "/timeseriesDefinitionByParameters",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<List<TimeSeriesHead>> getTimeseriesDefinitionByAccessParamterHeadDtos(@ApiParam(value = ""  )  @Valid @RequestBody List<AccessParamterHeadDto> accessParamtersHeadDto);


    @ApiOperation(value = "Create new TimeseriesDefinitions", nickname = "putTimeseriesDefinitionsByPersistenceParameterHeadDtos", notes = "", response = TimeSeriesHead.class, responseContainer = "List", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = TimeSeriesHead.class, responseContainer = "List"),
        @ApiResponse(code = 401, message = "unauthorized", response = Error.class),
        @ApiResponse(code = 404, message = "resource not found", response = Error.class),
        @ApiResponse(code = 200, message = "error", response = Error.class) })
    @RequestMapping(value = "/timeseriesDefinitionByParameters",
        produces = { "application/json" }, 
        method = RequestMethod.PUT)
    ResponseEntity<List<TimeSeriesHead>> putTimeseriesDefinitionsByPersistenceParameterHeadDtos(@ApiParam(value = ""  )  @Valid @RequestBody List<PersistenceParameterHeadDto> persistenceParametersHeadDto);

}